requirements.txt
flask==2.3.3
flask-cors==4.0.0
asana==3.2.1
pandas==2.1.1
python-dotenv==1.0.0
gunicorn==21.2.0 